Day2 apr24 pm

+++ app/Http/Controllers/ListingController.php
@@ -57,6 +57,11 @@ class ListingController extends Controller
   // Update Listing Data
   public function update(Request $request, Listing $listing)
   {
+    // Verify owner
+    if ($listing->user_id != auth()->id()) {
+      abort(403, 'Unauthorized Action');
+    }
+
     $formFields = $request->validate([
       'title' => 'required',
       'company' => ['required'],
@@ -85,7 +90,17 @@ class ListingController extends Controller
   // Delte Listing
   public function destroy(Listing $listing)
   {
+    if ($listing->user_id != auth()->id()) {
+      abort(403, 'Unauthorized Action');
+    }
+
     $listing->delete();
     return redirect('/')->with('message', 'Listing deleted duccessfully');
+  }
+
+  // Manage Listings
+  public function manage()
+  {
+    return view('listings.manage', ['listings' => auth()->user()->listings()->get()]);
   }
 }
